% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{inputs}
\alias{inputs}
\title{Add an Input to a Flow}
\usage{
inputs(flow, ...)
}
\arguments{
\item{flow}{(a flow) The flow where to add the input}

\item{...}{(list or vector) List of the inputs to add, can be given with or without quotes, e.g. list(A, B) or list("A", "B"). Default: list()}
}
\value{
Returns (invisibly) the flow with added inputs.
}
\description{
This function adds a node of type input to a flow.
}
