% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log}
\alias{log}
\title{Log of a Workflow}
\usage{
log(flow, level = c("DEBUG", "WARNING", "INFO", "ERROR"))
}
\arguments{
\item{flow}{The \code{workflow} to use}

\item{level}{Any of "DEBUG", "WARNING", "INFO" or "ERROR", indicating the level of verbosity of the log.}
}
\description{
Log of a Workflow
}
