% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_package.R
\name{pack}
\alias{pack}
\title{Create an R Package from a Flow}
\usage{
pack(flow, path, package_name = flow$name, open = FALSE)
}
\arguments{
\item{flow}{The flow to export as a package}

\item{path}{Root folder where to create the package}

\item{package_name}{Package name. The default is the name assgined to the workflow.}

\item{open}{Logical value indicating if the new project should be opened in a new RStudio window.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Create an R Package from a Flow
}
