% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Execute a Flow}
\usage{
run(
  flow,
  inputs = list(),
  desired_outputs = NULL,
  initialize_outputs = TRUE,
  cleanup = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{flow}{(a workflow) The flow to be executed.}

\item{inputs}{(list) Inputs to use, Default: list()}

\item{desired_outputs}{(character array) List of outputs to compute, Default: NULL}

\item{initialize_outputs}{(logical) initialize outputs?, Default: TRUE}

\item{cleanup}{(logical) perform cleanup of intermediate results?, Default: TRUE}

\item{verbose}{(logical) print information of the executions?, Default: FALSE}

\item{...}{Other arguments passed to subfunctions.}
}
\value{
A list with one (named) field for each \code{desired_output}.
}
\description{
This function allows to run a pipeline (flow) given some of its inputs.
}
