% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step}
\alias{step}
\title{Add a Function to a Flow}
\usage{
step(
  flow,
  f,
  inputs = ifelse(inherits(f, "function"), list(names(formals(f))), list()),
  output,
  ...
)
}
\arguments{
\item{flow}{(a flow) The flow where to add a new function}

\item{f}{(a function) The function itself to be added}

\item{inputs}{(list) List of the inputs needed for the function to be executed, defaults to the formal arguments of the function,
list())}

\item{output}{(name) The name to assign to the output of the function.}

\item{...}{extra parameters for the function, allows to parameterize the flow.}
}
\value{
The flow with the function added
}
\description{
This function adds another function to a given flow.
}
