% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyneuro-package.R
\docType{package}
\name{tidyneuro-package}
\alias{tidyneuro}
\alias{tidyneuro-package}
\title{tidyneuro: Tidy workflows for Neuroimaging}
\description{
Set up and execute neuroimaging pipelines and workflows with replicability, robustness and tidyness in mind.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/neuroimaginador/tidyneuro}
  \item Report bugs at \url{https://github.com/neuroimaginador/tidyneuro/issues}
}

}
\author{
\strong{Maintainer}: Domingo López \email{dominlopez@uma.es} (\href{https://orcid.org/0000-0002-0172-1585}{ORCID})

}
\keyword{internal}
