% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update a Workflow Node}
\usage{
update(flow, output, new_function)
}
\arguments{
\item{flow}{The \code{workflow} object to update}

\item{output}{The name of the output to update}

\item{new_function}{A function that will substitute the one already stored in the \code{output}}
}
\value{
Invisibly, the updated flow
}
\description{
Update a Workflow Node
}
