% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{workflow}
\alias{workflow}
\title{Create a Workflow}
\usage{
workflow(name = "", work_dir = tempdir(), inputs = list())
}
\arguments{
\item{name}{(string) The name given to the workflow}

\item{work_dir}{Working directory}

\item{inputs}{List of inputs}
}
\description{
Create a Workflow
}
