% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.workflow}
\alias{print.workflow}
\title{Print Flow Information}
\usage{
\method{print}{workflow}(flow)
}
\arguments{
\item{flow}{(a \code{workflow} object) Flow to print}
}
\value{
Prints the name, inputs, outputs, package dependencies and memory used.
}
\description{
Print Flow Information
}
