% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.workflow}
\alias{summary.workflow}
\title{Summary of Flow}
\usage{
\method{summary}{workflow}(object, ...)
}
\arguments{
\item{object}{(a \code{NIflow} object) Flow to print summary of.}

\item{...}{Not used.}
}
\value{
Prints inputs, outputs and dependencies between them.
}
\description{
Summary of Flow
}
